Initial implementation of `git-fixup` command
Creates a git fixup commit that is amended automatically to the target
commit.

'fzf' is used to improve the user experience by providing a selectable
list of usable commits. If 'fzf' is not found in the PATH user must
provide the commit hash to the 'git-fixup' command.

This command is intended to be used as a git alias. Aliases can be added
to '.gitconfig' file. Add an alias to the '.gitconfig' file called
'fixup' (or some other preferred name) and point it to the location of
this script.

```.gitconfig
[alias]
        fixup = git-fixup
```

USAGE:

```sh
git-fixup [<commit_hash>]
```

When commit hash is not provided `git-fixup` will use 'fzf' to provide a list
of selectable commits in the branch, however, if 'fzf' is not found in the
PATH, user _must_ provide a commit hash.